+++ lib/npm.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-npm_version_major() {
-  npm --version | cut -d "." -f 1
-}